+++ app/Filament/Resources/OfferwallProviderSettingResource/Pages/CreateOfferwallProviderSetting.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Filament\Resources\OfferwallProviderSettingResource\Pages;
+
+use App\Filament\Resources\OfferwallProviderSettingResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateOfferwallProviderSetting extends CreateRecord
+{
+    protected static string $resource = OfferwallProviderSettingResource::class;
+}